Extract FileContext from lib
