+++ src/lib.rs
@@ -40,7 +40,7 @@ extern crate num_rational;
 extern crate num_traits;
 extern crate rand;
 
-use std::path::{Path, PathBuf};
+use std::path::Path;
 
 mod colors;
 mod error;
@@ -52,8 +52,11 @@ mod variablescope;
 mod output_style;
 mod unit;
 mod parser;
+mod file_context;
 
 pub use error::Error;
+
+pub use file_context::FileContext;
 use formalargs::{CallArgs, FormalArgs};
 pub use functions::SassFunction;
 pub use num_rational::Rational;
@@ -125,66 +128,6 @@ pub fn compile_scss_file(file: &Path,
     let (sub_context, file) = file_context.file(file);
     let items = parse_scss_file(&file)?;
     style.write_root(&items, &mut GlobalScope::new(), sub_context)
-}
-
-/// A file context specifies where to find files to load.
-///
-/// When opening an included file, an extended file context is
-/// created, to find further included files relative to the file they
-/// are inlcuded from.
-///
-/// # Example
-/// ```
-/// use rsass::FileContext;
-///
-/// let base = FileContext::new();
-/// let (base, file1) = base.file("some/dir/file.scss".as_ref());
-/// // base is now a relative to file1, usefull to open files
-/// // by paths mentioned in file1.
-/// let (base, file2) = base.file("some/other.scss".as_ref());
-/// assert_eq!(file1.to_string_lossy(), "some/dir/file.scss");
-/// assert_eq!(file2.to_string_lossy(), "some/dir/some/other.scss");
-/// ```
-#[derive(Clone, Debug)]
-pub struct FileContext {
-    path: PathBuf,
-}
-
-impl FileContext {
-    /// Create a new FileContext.
-    ///
-    /// Files will be resolved from the current working directory.
-    pub fn new() -> Self {
-        FileContext { path: PathBuf::new() }
-    }
-    /// Get a file from this context.
-    ///
-    /// Get a path and a FileContext from this FileContext and a path.
-    pub fn file(&self, file: &Path) -> (Self, PathBuf) {
-        let t = self.path.join(file);
-        if let Some(dir) = t.parent() {
-            (FileContext { path: PathBuf::from(dir) }, t.clone())
-        } else {
-            (FileContext::new(), t.clone())
-        }
-    }
-    fn find_file(&self, name: &Path) -> Option<(Self, PathBuf)> {
-        // TODO Check docs what expansions should be tried!
-        let parent = name.parent();
-        if let Some(name) = name.file_name().and_then(|n| n.to_str()) {
-            for name in &[name,
-                          &format!("{}.scss", name),
-                          &format!("_{}.scss", name)] {
-                let full =
-                    parent.map(|p| p.join(name)).unwrap_or_else(|| name.into());
-                let (c, p) = self.file(&full);
-                if p.is_file() {
-                    return Some((c, p));
-                }
-            }
-        }
-        None
-    }
 }
 
 /// Every sass file is a sequence of sass items.
